+++ src/main/kotlin/com/ligon/deeznuts/services/UserService.kt
@@ -4,6 +4,6 @@ import com.ligon.deeznuts.exceptions.NotFoundException
 import com.ligon.deeznuts.repositories.UserRepository
 
 class UserService(private val userRepository: UserRepository) {
-    suspend fun findById(id: Int) = userRepository.findById(id) ?: throw NotFoundException()
+    suspend fun findById(id: Long) = userRepository.findById(id) ?: throw NotFoundException()
 }
